fix:Show Selected Event Type and Convert the start time and end time to UTC.
