+++ apps/front-end/public/locales/en/translation.json
@@ -523,6 +523,7 @@
   "EVENT_END_DATE_TIME": "End Date & Time",
   "EVENT_NAME": "Event Name \n (Batch Number)",
   "EVENT_START_DATE_TIME": "Start Date & Time",
+  "EVENT_SCHEDULED": "Event Scheduled",
   "EVENT_TYPE": "Event Type",
   "EXAM_ALREADY_TAKEN": "Exam already attempted",
   "EXAM_SUBMITTED_SUCCESSFULLY": "Test has been submitted successfully",
@@ -540,6 +541,7 @@
   "FACILITATOR_EVENT_TYPE_PRAGATI_ORIENTATION": "Pragati Orientation",
   "FACILITATOR_EVENT_TYPE_PRERAK_CAMP_EXECUTION_TRAINING": "Prerak Camp Execution Training",
   "FACILITATOR_EVENT_TYPE_PRERAK_FLN_TRAINING": "Prerak FLN Training",
+  "FACILITATOR_EVENT_TYPE_PRERAK_ORIENTATION": "Prerak Orientation",
   "FACILITATOR_FULL_TIME": "Full Time",
   "FACILITATOR_MARITAL_STATUS_DIVORCED": "Divorced",
   "FACILITATOR_MARITAL_STATUS_MARRIED": "Married",
