+++ apps/front-end/public/locales/hi/translation.json
@@ -524,6 +524,7 @@
   "EVENT_END_DATE_TIME": "समाप्ति की तारीख और समय",
   "EVENT_NAME": "कार्यक्रम का नाम \n(बैच संख्या)",
   "EVENT_START_DATE_TIME": "शुरुआत की तारीख और समय",
+  "EVENT_SCHEDULED": "कार्यक्रम निर्धारित किया गया है",
   "EVENT_TYPE": "कार्यक्रम का प्रकार",
   "EXAM_ALREADY_TAKEN": "परीक्षा का प्रयास पहले ही किया जा चुका है",
   "EXAM_SUBMITTED_SUCCESSFULLY": "परीक्षण सफलतापूर्वक सबमिट कर दिया गया है",
@@ -541,6 +542,7 @@
   "FACILITATOR_EVENT_TYPE_PRAGATI_ORIENTATION": "प्रगति ओरिएंटेशन",
   "FACILITATOR_EVENT_TYPE_PRERAK_CAMP_EXECUTION_TRAINING": "प्रेरक शिविर क्रियान्वयन प्रशिक्षण",
   "FACILITATOR_EVENT_TYPE_PRERAK_FLN_TRAINING": "प्रेरक FLN प्रशिक्षण",
+  "FACILITATOR_EVENT_TYPE_PRERAK_ORIENTATION": "प्रेरक ओरिएंटेशन",
   "FACILITATOR_FULL_TIME": "फुल टाइम",
   "FACILITATOR_MARITAL_STATUS_DIVORCED": "तलाकशुदा",
   "FACILITATOR_MARITAL_STATUS_MARRIED": "शादीशुदा",
