+++ apps/front-end/src/pages/admin/event/EventFrom.js
@@ -1,29 +1,23 @@
-import React, { useCallback, useEffect, useRef, useState } from "react";
-import PropTypes from "prop-types";
+import Form from "@rjsf/core";
+import validator from "@rjsf/validator-ajv8";
 import {
+  AdminTypo,
+  BodyMedium,
   IconByName,
   AdminLayout as Layout,
-  AdminTypo,
-  campService,
-  CardComponent,
-  UserCard,
-  MapComponent,
-  CheckUncheck,
-  enumRegistryService,
-  jsonParse,
-  ImageView,
-  BodyMedium,
-  GetEnumValue,
-  mapDistance,
-  geolocationRegistryService,
-  facilitatorRegistryService,
-  getOptions,
-  eventService,
   arrList,
+  enumRegistryService,
+  eventService,
+  facilitatorRegistryService,
+  geolocationRegistryService,
+  getOptions,
+  jsonParse,
 } from "@shiksha/common-lib";
+import { ChipStatus } from "component/Chip";
+import { debounce } from "lodash";
+import moment from "moment";
 import {
   Alert,
-  Box,
   Checkbox,
   HStack,
   Input,
@@ -33,16 +27,13 @@ import {
   VStack,
   useToast,
 } from "native-base";
-import { useTranslation } from "react-i18next";
-import Form from "@rjsf/core";
-import Schema from "./Schema";
-import validator from "@rjsf/validator-ajv8";
-import { widgets, templates } from "../../../component/BaseInput";
-import moment from "moment";
+import PropTypes from "prop-types";
+import { useCallback, useEffect, useRef, useState } from "react";
 import DataTable from "react-data-table-component";
-import { ChipStatus } from "component/Chip";
-import { debounce } from "lodash";
+import { useTranslation } from "react-i18next";
 import { useNavigate, useParams } from "react-router-dom";
+import { templates, widgets } from "../../../component/BaseInput";
+import Schema from "./Schema";
 
 const customStyles = {
   headCells: {
