--- apps/front-end/src/pages/front-end/Attendence/schema.js
@@ -1,15 +1,0 @@
-export default {
-  type: "object",
-  properties: {
-    documents_status: {
-      label: "Reminder",
-      type: "array",
-      title: "DOCUMENT VERIFICATION",
-      items: {
-        type: "string",
-        enum: ["Qualification Certificate", "Volunteer Proof", "Work Proof"],
-      },
-      uniqueItems: true,
-    },
-  },
-};
